fix(edit): close certified projection review gaps
