+++ tests/unit/test_symbols_json_enrichment.py
@@ -14,8 +14,20 @@ both weights. Only the serialization half ships; pilot data on #517.
 
 from __future__ import annotations
 
+import ast
+from types import SimpleNamespace
+
 import pytest
 
+from tree_sitter_analyzer.cache import _symbol_walker as walker_module
+from tree_sitter_analyzer.cache._symbol_walker import (
+    _cpp_code_mask,
+    _python_dynamic_loader_analysis,
+    _python_module_control_bindings,
+    _python_module_scope_statements,
+    _SymbolWalker,
+    _walk_for_symbols,
+)
 from tree_sitter_analyzer.cache.extraction import (
     _DOCSTRING_MAX_CHARS,
     _extract_imports,
@@ -155,9 +167,63 @@ class TestExtractorVersionBump:
         from tree_sitter_analyzer import ast_cache
         from tree_sitter_analyzer.cache import indexer as _ast_cache_indexer
 
-        # v24: CommonJS aliases and C++20 imports enter the projection.
-        assert ast_cache._AST_CACHE_EXTRACTOR_VERSION == 24
-        assert _ast_cache_indexer._AST_CACHE_EXTRACTOR_VERSION == 24
+        # v25: parse-error state and current projection semantics are persisted.
+        assert ast_cache._AST_CACHE_EXTRACTOR_VERSION == 25
+        assert _ast_cache_indexer._AST_CACHE_EXTRACTOR_VERSION == 25
+
+
+def test_python_module_control_bindings_cover_all_module_control_targets() -> None:
+    """Control-flow rebinding is complete without descending into nested scopes."""
+    module = ast.parse(
+        """
+async def nested_async():
+    hidden_async = 1
+
+class Nested:
+    hidden_class = 1
+
+lambda: hidden_lambda
+[hidden_list for hidden_list in values]
+{hidden_set for hidden_set in values}
+{hidden_key: hidden_value for hidden_key, hidden_value in pairs}
+(hidden_generator for hidden_generator in values)
+
+async for async_item in items:
+    pass
+
+async with manager(), other() as async_bound:
+    pass
+
+with manager(), other() as sync_bound:
+    pass
+
+try:
+    pass
+except Exception:
+    pass
+
+match value:
+    case [*rest]:
+        pass
+    case [*_]:
+        pass
+    case {"key": item, **remaining}:
+        pass
+    case {}:
+        pass
+    case _:
+        pass
+"""
+    )
+
+    assert _python_module_control_bindings(module.body) == {
+        "async_item",
+        "async_bound",
+        "sync_bound",
+        "rest",
+        "item",
+        "remaining",
+    }
 
 
 class TestJavaScriptModuleCallProjection:
@@ -262,6 +328,36 @@ class TestJavaScriptModuleCallProjection:
         symbols = {"symbols": _symbols_for(source, "typescript")}
 
         assert _extract_imports(symbols) == [{"text": "load('./legacy')", "line": 2}]
+
+    def test_comment_text_does_not_create_commonjs_loader_alias(self) -> None:
+        source = (
+            "// const load = require\n"
+            "function load(value) { return value; }\n"
+            "load('./legacy');"
+        )
+        extraction = _extraction_for(source, "typescript")
+
+        assert _extract_imports(extraction) == []
+        assert extraction["import_projection_complete"] is True
+
+    def test_commonjs_loader_rebound_by_loop_marks_projection_incomplete(self) -> None:
+        extraction = _extraction_for(
+            "for (require of loaders) { require('./legacy'); }", "typescript"
+        )
+
+        assert extraction["import_projection_complete"] is False
+
+    def test_syntax_error_is_persisted_with_projection(self) -> None:
+        from tree_sitter_analyzer.cache.extraction import _extract_symbols
+        from tree_sitter_analyzer.core.parser import Parser
+
+        source = "if ( {\nrequire('./legacy');"
+        result = Parser().parse_code(source, "typescript")
+        assert result.tree is not None
+
+        extraction = _extract_symbols(result.tree, source, "typescript")
+
+        assert extraction["syntax_error"] is True
 
     def test_incomplete_call_node_is_not_projected_as_import(self):
         from tree_sitter_analyzer.cache._symbol_walker import _SymbolWalker
@@ -630,3 +726,574 @@ class TestCodexP2sOn621:
         src = "function f(): string { return 'x'; }\n"
         syms = {x["name"]: x for x in _symbols_for(src, "typescript")}
         assert syms["f"]["return_type"] == "string"
+
+
+@pytest.mark.parametrize(
+    ("source", "visible"),
+    [
+        ("// comment", ""),
+        ("// comment\nx", "\nx"),
+        ('R"(unterminated', ""),
+        ('R"12345678901234567(payload)', "R"),
+        ('"a\\"b"x', "x"),
+    ],
+)
+def test_cpp_code_mask_excludes_comments_and_literals(
+    source: str, visible: str
+) -> None:
+    mask = _cpp_code_mask(source)
+
+    assert (
+        "".join(char for char, is_code in zip(source, mask, strict=True) if is_code)
+        == visible
+    )
+
+
+def test_cpp_module_fallback_does_not_duplicate_walker_projection(monkeypatch) -> None:
+    source = "import project.core;"
+
+    def project_import(walker: _SymbolWalker, *_args: object) -> None:
+        walker.symbols.append(
+            {
+                "kind": "import",
+                "text": source,
+                "line": 1,
+                "language": "cpp",
+            }
+        )
+
+    monkeypatch.setattr(_SymbolWalker, "walk", project_import)
+    tree = SimpleNamespace(root_node=SimpleNamespace(children=[]))
+
+    result = walker_module._extract_symbols(tree, source, "cpp")
+
+    assert result["symbols"] == [
+        {
+            "kind": "import",
+            "text": source,
+            "line": 1,
+            "language": "cpp",
+        }
+    ]
+
+
+def test_python_loader_analysis_fails_closed_on_invalid_module() -> None:
+    names, complete = _python_dynamic_loader_analysis("if (")
+
+    assert names == frozenset(
+        {"__import__", "builtins.__import__", "importlib.import_module"}
+    )
+    assert complete is False
+
+
+def test_python_loader_analysis_tracks_module_alias_chains() -> None:
+    source = """
+import importlib as il
+from importlib import import_module, invalidate_caches
+loader: object = il.import_module
+later = loader
+annotation_only: object
+"""
+
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert names == frozenset(
+        {
+            "__import__",
+            "builtins.__import__",
+            "importlib.import_module",
+            "il.import_module",
+            "import_module",
+            "loader",
+            "later",
+        }
+    )
+    assert complete is True
+
+
+def test_python_loader_analysis_rejects_nested_bindings_and_aliases() -> None:
+    source = """
+import importlib
+
+def load(positional, /, regular, *items, keyword, **options):
+    nested = importlib.import_module
+    holder.loader = importlib.import_module
+    annotated: object = importlib.import_module
+    import importlib as nested_importlib
+    from importlib import import_module as nested_load
+"""
+
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert "importlib.import_module" in names
+    assert complete is False
+
+
+@pytest.mark.parametrize(
+    "body",
+    [
+        "for loader in values:\n        loader('pkg.util')",
+        "try:\n        pass\n    except Exception as loader:\n        loader('pkg.util')",
+        "if (nested := importlib.import_module):\n        nested('pkg.util')",
+    ],
+)
+def test_python_loader_analysis_rejects_additional_lexical_bindings(
+    body: str,
+) -> None:
+    source = (
+        "import importlib\n"
+        "loader = importlib.import_module\n"
+        "def load(values):\n"
+        f"    {body}\n"
+    )
+
+    _names, complete = _python_dynamic_loader_analysis(source)
+
+    assert complete is False
+
+
+def test_python_loader_analysis_discovers_module_control_flow_alias() -> None:
+    source = """
+def load_plugin():
+    return load("pkg.util")
+
+if enabled:
+    from importlib import import_module as load
+"""
+
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert "load" in names
+    assert complete is True
+
+
+@pytest.mark.parametrize(
+    "control_flow",
+    [
+        "for load in loaders:\n    load('pkg.util')",
+        "with manager() as load:\n    load('pkg.util')",
+        "try:\n    pass\nexcept Exception as load:\n    load('pkg.util')",
+        "if (load := fake):\n    load('pkg.util')",
+        "match value:\n    case load:\n        load('pkg.util')",
+    ],
+)
+def test_python_loader_analysis_rejects_module_control_flow_rebinding(
+    control_flow: str,
+) -> None:
+    source = f"from importlib import import_module as load\n{control_flow}\n"
+
+    _names, complete = _python_dynamic_loader_analysis(source)
+
+    assert complete is False
+
+
+def test_python_loader_analysis_rejects_module_rebinding() -> None:
+    source = """
+from importlib import import_module as load
+load = fake
+load("pkg.util")
+"""
+
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert "load" in names
+    assert complete is False
+
+
+@pytest.mark.parametrize(
+    "source",
+    [
+        "import importlib\nimportlib.import_module = fake\n"
+        "importlib.import_module('pkg.util')\n",
+        "import builtins as runtime\nruntime.__import__ = fake\n"
+        "runtime.__import__('pkg.util')\n",
+        "import importlib\nimportlib.import_module, other = fake, value\n"
+        "importlib.import_module('pkg.util')\n",
+        "import builtins as runtime\n[other, [runtime.__import__]] = values\n"
+        "runtime.__import__('pkg.util')\n",
+        "import builtins as runtime\nother, *runtime.__import__ = values\n"
+        "runtime.__import__('pkg.util')\n",
+        "import importlib\nload, other = importlib.import_module, value\n"
+        "load('pkg.util')\n",
+        "import importlib\npair = (importlib.import_module, value)\n"
+        "load, other = pair\nload('pkg.util')\n",
+        "import importlib\npair = {'load': importlib.import_module}\n"
+        "load = pair['load']\nload('pkg.util')\n",
+        "import importlib\npair = wrap(importlib.import_module)\n"
+        "load = pair[0]\nload('pkg.util')\n",
+        "import importlib\nholder.load = importlib.import_module\n"
+        "holder.load('pkg.util')\n",
+        "import importlib\nregistry['load'] = importlib.import_module\n"
+        "registry['load']('pkg.util')\n",
+    ],
+)
+def test_python_loader_analysis_rejects_qualified_loader_rebinding(
+    source: str,
+) -> None:
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert any(name.endswith((".import_module", ".__import__")) for name in names)
+    assert complete is False
+
+
+@pytest.mark.parametrize(
+    "source",
+    [
+        "import importlib\nloaded = importlib.import_module('pkg.util')\n",
+        "import importlib\ndef load():\n"
+        "    loaded = importlib.import_module('pkg.util')\n"
+        "    return loaded\n",
+    ],
+)
+def test_python_loader_analysis_accepts_assigned_loader_call_result(
+    source: str,
+) -> None:
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert "importlib.import_module" in names
+    assert complete is True
+
+
+@pytest.mark.parametrize(
+    "source",
+    [
+        "import importlib\ndef load(callback=importlib.import_module):\n"
+        "    return callback('pkg.util')\n",
+        "import importlib\ndef load(value):\n"
+        "    pair = (importlib.import_module, value)\n"
+        "    callback, other = pair\n"
+        "    return callback('pkg.util')\n",
+        "import importlib\nload = lambda callback=importlib.import_module: "
+        "callback('pkg.util')\n",
+    ],
+)
+def test_python_loader_analysis_rejects_nested_stored_loader(source: str) -> None:
+    _names, complete = _python_dynamic_loader_analysis(source)
+
+    assert complete is False
+
+
+@pytest.mark.parametrize(
+    "source",
+    [
+        "import importlib\ndef holder(value: importlib.import_module):\n    pass\n",
+        "import importlib\ndef holder(*values: importlib.import_module):\n    pass\n",
+        "import importlib\ndef holder(**values: importlib.import_module):\n    pass\n",
+        "import importlib\ndef holder() -> importlib.import_module:\n    pass\n",
+        "import importlib\nholder: importlib.import_module\n",
+        "import importlib\ndef outer():\n    holder: importlib.import_module\n",
+    ],
+)
+def test_python_loader_analysis_rejects_annotation_storage(source: str) -> None:
+    _names, complete = _python_dynamic_loader_analysis(source)
+
+    assert complete is False
+
+
+@pytest.mark.parametrize(
+    "rebind",
+    [
+        "from helpers import load",
+        "def load():\n    pass",
+        "class load:\n    pass",
+        "del load",
+    ],
+)
+def test_python_loader_analysis_rejects_nonassignment_rebinding(
+    rebind: str,
+) -> None:
+    source = (
+        f"from importlib import import_module as load\n{rebind}\nload('pkg.util')\n"
+    )
+
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert "load" in names
+    assert complete is False
+
+
+def test_python_loader_analysis_tracks_builtins_alias() -> None:
+    names, complete = _python_dynamic_loader_analysis(
+        "import builtins as runtime\nruntime.__import__('pkg.util')\n"
+    )
+
+    assert "runtime.__import__" in names
+    assert complete is True
+
+
+@pytest.mark.parametrize(
+    "expression",
+    [
+        "callback = lambda load: load('pkg.util')",
+        "results = [load('pkg.util') for load in loaders]",
+    ],
+)
+def test_python_loader_analysis_rejects_expression_scope_shadowing(
+    expression: str,
+) -> None:
+    source = f"from importlib import import_module as load\n{expression}\n"
+
+    names, complete = _python_dynamic_loader_analysis(source)
+
+    assert "load" in names
+    assert complete is False
+
+
+def test_module_scope_statement_walk_skips_nested_function_body() -> None:
+    module = ast.parse("if enabled:\n    def nested():\n        import importlib\n")
+
+    statements = _python_module_scope_statements(module)
+
+    assert [type(statement) for statement in statements] == [ast.If, ast.FunctionDef]
+
+
+@pytest.mark.parametrize(
+    ("language", "node_type", "enclosed"),
+    [
+        ("python", "class_definition", False),
+        ("scala", "identifier", False),
+        ("scala", "object_definition", True),
+    ],
+)
+def test_scala_projection_rejects_non_top_level_class_like_nodes(
+    language: str, node_type: str, enclosed: bool
+) -> None:
+    walker = _SymbolWalker("", [], language, None)
+    node = SimpleNamespace(type=node_type)
+
+    assert walker._append_scala(node, enclosed) is False
+
+
+def test_scala_projection_handles_empty_and_present_symbols(monkeypatch) -> None:
+    walker = _SymbolWalker("", [], "scala", None)
+    node = SimpleNamespace(type="object_definition")
+    monkeypatch.setattr(walker_module, "_scala_symbol_from_node", lambda *_: None)
+
+    assert walker._append_scala(node, False) is True
+    assert walker.symbols == []
+
+    expected = {"kind": "class", "name": "User"}
+    monkeypatch.setattr(walker_module, "_scala_symbol_from_node", lambda *_: expected)
+    assert walker._append_scala(node, False) is True
+    assert walker.symbols == [expected]
+
+
+def test_collect_node_stops_after_scala_projection(monkeypatch) -> None:
+    expected = {"kind": "object", "name": "Registry"}
+    monkeypatch.setattr(walker_module, "_scala_symbol_from_node", lambda *_: expected)
+    node = SimpleNamespace(
+        type="object_definition", child_by_field_name=lambda _name: None
+    )
+    walker = _SymbolWalker("", [], "scala", None)
+
+    walker._collect_node(node, 0, False)
+
+    assert walker.symbols == [expected]
+
+
+class _TextNode:
+    def __init__(self, node_type: str, source: str) -> None:
+        self.type = node_type
+        self.start_byte = 0
+        self.end_byte = len(source.encode())
+        self.start_point = (0, 0)
+        self.end_point = (0, len(source))
+
+
+def test_python_loader_assignment_rejects_invalid_syntax() -> None:
+    source = "if ("
+    walker = _SymbolWalker(source, [], "python", None)
+
+    assert (
+        walker._append_python_loader_assignment(_TextNode("assignment", source))
+        is False
+    )
+
+
+def test_python_loader_assignment_rejects_multiple_statements() -> None:
+    source = "first = importlib.import_module\nsecond = first"
+    walker = _SymbolWalker(source, [], "python", None)
+
+    assert (
+        walker._append_python_loader_assignment(_TextNode("assignment", source))
+        is False
+    )
+
+
+def test_python_loader_assignment_accepts_annotated_alias() -> None:
+    source = "loader: object = importlib.import_module"
+    walker = _SymbolWalker(source, [], "python", None)
+
+    assert (
+        walker._append_python_loader_assignment(
+            _TextNode("annotated_assignment", source)
+        )
+        is True
+    )
+
+
+def test_python_loader_assignment_rejects_annotation_without_value() -> None:
+    source = "loader: object"
+    walker = _SymbolWalker(source, [], "python", None)
+
+    assert (
+        walker._append_python_loader_assignment(
+            _TextNode("annotated_assignment", source)
+        )
+        is False
+    )
+
+
+class _FieldNode(_TextNode):
+    def __init__(
+        self, node_type: str, source: str, fields: dict[str, _TextNode]
+    ) -> None:
+        super().__init__(node_type, source)
+        self.fields = fields
+
+    def child_by_field_name(self, name: str):
+        return self.fields.get(name)
+
+
+def test_java_reflection_rejects_missing_name() -> None:
+    source = "forName()"
+    node = _FieldNode(
+        "method_invocation", source, {"arguments": _TextNode("arguments", source)}
+    )
+
+    assert (
+        _SymbolWalker(source, [], "java", None)._append_java_reflective_load(node)
+        is False
+    )
+
+
+def test_java_reflection_rejects_other_method_name() -> None:
+    source = "load()"
+    node = _FieldNode(
+        "method_invocation",
+        source,
+        {
+            "name": _TextNode("identifier", "load"),
+            "arguments": _TextNode("arguments", source),
+        },
+    )
+
+    assert (
+        _SymbolWalker(source, [], "java", None)._append_java_reflective_load(node)
+        is False
+    )
+
+
+def test_include_next_projection_ignores_other_preprocessor_calls() -> None:
+    source = "#pragma once"
+    node = _TextNode("preproc_call", source)
+
+    assert _SymbolWalker(source, [], "cpp", None)._append_include_next(node) is False
+
+
+def test_historical_walk_wrapper_delegates_to_walker() -> None:
+    node = SimpleNamespace(
+        type="comment", children=[], child_by_field_name=lambda _name: None
+    )
+    symbols: list[dict] = []
+
+    _walk_for_symbols(node, "", symbols, "unknown")
+
+    assert symbols == []
+
+
+def test_symbol_walk_records_depth_truncation() -> None:
+    truncated = [False]
+    walker = _SymbolWalker("", [], "unknown", truncated)
+
+    walker.walk(object(), depth=10_000)
+
+    assert truncated == [True]
+
+
+def test_symbol_walk_without_truncation_sink_stops_cleanly() -> None:
+    walker = _SymbolWalker("", [], "unknown", None)
+
+    walker.walk(object(), depth=10_000)
+
+    assert walker.symbols == []
+
+
+def test_php_constant_projection_appends_extracted_constants(monkeypatch) -> None:
+    expected = {"kind": "constant", "name": "LIMIT"}
+    monkeypatch.setattr(walker_module, "_php_constants", lambda *_: [expected])
+    node = SimpleNamespace(type="const_declaration")
+    walker = _SymbolWalker("", [], "php", None)
+
+    walker._append_constant(node, None, False)
+
+    assert walker.symbols == [expected]
+
+
+def test_go_constant_projection_appends_extracted_constants(monkeypatch) -> None:
+    expected = {"kind": "constant", "name": "Limit"}
+    monkeypatch.setattr(walker_module, "_go_package_constants", lambda *_: [expected])
+    node = SimpleNamespace(type="const_declaration")
+    walker = _SymbolWalker("", [], "go", None)
+
+    walker._append_constant(node, None, False)
+
+    assert walker.symbols == [expected]
+
+
+def test_rust_constant_projection_appends_named_constant() -> None:
+    source = "LIMIT"
+    name = _TextNode("identifier", source)
+    node = SimpleNamespace(
+        type="const_item", start_point=(0, 0), end_point=(0, len(source))
+    )
+    walker = _SymbolWalker(source, [], "rust", None)
+
+    walker._append_constant(node, name, False)
+
+    assert walker.symbols == [
+        {
+            "kind": "constant",
+            "name": "LIMIT",
+            "line": 1,
+            "end_line": 1,
+            "language": "rust",
+        }
+    ]
+
+
+def test_class_projection_covers_empty_fallback_and_parents(monkeypatch) -> None:
+    empty_name = SimpleNamespace(type="identifier", start_byte=0, end_byte=0)
+    node = SimpleNamespace(
+        type="class_definition",
+        children=[SimpleNamespace(type="comment"), empty_name],
+        start_point=(0, 0),
+        end_point=(0, 0),
+    )
+    walker = _SymbolWalker("", [], "python", None)
+
+    assert walker._class_name_node(node, object()) is not None
+    assert walker._class_name_node(node, None) is empty_name
+    assert (
+        walker._class_name_node(
+            SimpleNamespace(children=[SimpleNamespace(type="comment")]), None
+        )
+        is None
+    )
+    walker._append_class(node, None)
+    assert walker.symbols == []
+
+    monkeypatch.setattr(walker_module, "_node_text", lambda *_: "Child")
+    monkeypatch.setattr(walker_module, "_extract_parent_classes", lambda *_: ["Base"])
+    monkeypatch.setattr(walker_module, "_python_docstring", lambda *_: None)
+    walker._append_class(node, None)
+    assert walker.symbols == [
+        {
+            "kind": "class",
+            "name": "Child",
+            "line": 1,
+            "end_line": 1,
+            "language": "python",
+            "parents": ["Base"],
+        }
+    ]
